requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
aiohttp==3.9.0
tenacity>=8.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.20.0
